+++ src/components/bot/validators.ts
@@ -1,0 +1,18 @@
+import { checkSchema } from "express-validator";
+
+export const validateQrName = () =>
+  checkSchema({
+    name: {
+      in: ["params"],
+      isString: true,
+      isEmpty: false,
+      trim: true,
+      optional: true,
+      isLength: {
+        options: { min: 1, max: 10 },
+      },
+      customSanitizer: {
+        options: (value: string) => value.toLowerCase().replaceAll(/[./]/g, ""),
+      },
+    },
+  });